% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LETTER.R
\docType{data}
\name{LETTER}
\alias{LETTER}
\title{Database of character image features.}
\format{A data frame with 20000 rows and 17 variables:
\describe{
  \item{class}{class labels}
  \item{V1}{first extracted feature}
  \item{V2}{second extracted feature}
  \item{V3}{third extracted feature}
  \item{V4}{4th extracted feature}
  \item{V5}{5th extracted feature}
  \item{V6}{6th extracted feature}
  \item{V7}{7th extracted feature}
  \item{V8}{8th extracted feature}
  \item{V9}{9th extracted feature}
  \item{V10}{10th extracted feature}
  \item{V11}{11th extracted feature}
  \item{V12}{12th extracted feature}
  \item{V13}{13th extracted feature}
  \item{V14}{14th extracted feature}
  \item{V15}{15th extracted feature}
  \item{V16}{16th extracted feature}
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/letter+recognition/}
}
\usage{
LETTER
}
\description{
A dataset containing 16 features extracted from 20000 handwritten characters.
}
\keyword{datasets}
